Workaround: Delete Raw Recordings
Despite not actually starting a recording, BigBlueButton still writes a
recording to disk. It will try to clean these recordings up in a regular
interval but unfortunately seems to forget half of the data.

This patch contains two Systemd series and timers which will:

- delete all raw recording every night at 4am
- reboot the server at 4:30am (shouldn't hurt and may help…)

This is a short-term workaround for #2. Nevertheless, this still needs a
proper solution.